refactor(app): 抽出 RemoteWorkflowOrchestrator，facade 退化为标准操作编排
Why:
- facade(LolAudioUnpackApp) 同时承担 update/extract/mapping/transcode 标准操作与整套 remote
  按实体拆批工作流(build_work_items/run_workflow/重试/清理)，787 行职责过载。

What:
- 新增 app/remote_workflow.py：RemoteWorkflowOrchestrator 专责 remote 实体工作流；标准操作、
  实体解析、读取器与远端准备器均经回调注入，避免反向依赖 facade。
- facade 移除 6 个远端私有方法与 build_work_items/run_workflow 主体，保留同名薄委托方法
  (经 _remote_orchestrator 构造编排器)；facade 由 787 行降至约 430 行。

Risk:
- 兼容性影响：无。对外 app.run_workflow/build_work_items 签名与行为保持，CLI(dispatch) 入口不变。
- 潜在风险：无明显额外风险（行为保持，已全量回归）。

Test:
- uv run pytest -q：593 passed, 4 deselected（remote_live）。测试仅最小改动：run_workflow 完成
  汇总日志移到新模块，test_remote_preparer 对应 logger patch 目标改为 m_remote_workflow 并加导入。
